requirements.txt
filelock==3.16.1
fsspec==2024.10.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
pillow==11.0.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
